allow to omit the default color palette
