+++ .changeset/breezy-beans-promise.md
@@ -1,0 +1,5 @@
+---
+'@twind/preset-tailwind': minor
+---
+
+allow to omit the default color palette
