+++ packages/preset-tailwind/base.d.ts
@@ -1,0 +1,4 @@
+// ONLY FOR TYPESCRIPT TO FIND @twind/preset-tailwind/base
+
+export { default } from './src/base'
+export * from './src/base'
