+++ packages/preset-tailwind/baseTheme.d.ts
@@ -1,0 +1,4 @@
+// ONLY FOR TYPESCRIPT TO FIND @twind/preset-tailwind/baseTheme
+
+export { default } from './src/baseTheme'
+export * from './src/baseTheme'
