+++ packages/preset-tailwind/index.d.ts
@@ -3,6 +3,8 @@
 export { default } from './src/index'
 export * from './src/index'
 
+export * as base from './src/base'
+export * as baseTheme from './src/baseTheme'
 export * as colors from './src/colors'
 export * as defaultTheme from './src/defaultTheme'
 export * as rules from './src/rules'
