+++ packages/preset-tailwind/src/index.ts
@@ -1,10 +1,8 @@
 import type { Preset } from '@twind/core'
 import type { TailwindTheme } from './types'
 
-import theme from './defaultTheme'
-import preflight from './preflight'
-import rules from './rules'
-import variants from './variants'
+import presetTailwindBase from './base'
+import * as colors from './colors'
 
 export * from './types'
 
@@ -16,19 +14,5 @@ export interface TailwindPresetOptions {
 export default function presetTailwind({
   disablePreflight,
 }: TailwindPresetOptions = {}): Preset<TailwindTheme> {
-  return {
-    // allow other preflight to run
-    preflight: disablePreflight ? undefined : preflight,
-    theme,
-    variants,
-    rules,
-    finalize(rule) {
-      // automatically add `content: ''` to before and after so you don’t have to specify it unless you want a different value
-      if (rule.r.some((r) => /^&::(before|after)$/.test(r)) && !rule.d?.includes('content:')) {
-        return { ...rule, d: ['content:var(--tw-content)', rule.d].filter(Boolean).join(';') }
-      }
-
-      return rule
-    },
-  }
+  return presetTailwindBase({ colors, disablePreflight })
 }
